pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimic-icd-coder"
version = "0.1.0"
description = "Multi-label ICD-10 auto-coding on MIMIC-IV discharge summaries"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Nancy Tanaka", email = "nancytanaka1@gmail.com" }]

dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pyarrow>=15.0",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "click>=8.1",
    "mlflow>=2.11",
    "joblib>=1.3",
]

[project.optional-dependencies]
transformer = [
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.28",
    "datasets>=2.18",
    "evaluate>=0.4",
    "sentencepiece>=0.2",
]
databricks = [
    "databricks-sdk>=0.22",
    "databricks-connect>=15.0",
]
serve = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
]
monitor = [
    "evidently>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "black>=24.3",
    "mypy>=1.9",
    "pandas-stubs",
    "types-PyYAML",
    "ipykernel>=6.29",
    "pre-commit>=3.7",
]

[project.scripts]
mic = "mimic_icd_coder.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers --cov=mimic_icd_coder --cov-report=term-missing"
